backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
pandas>=2.2.0
reportlab==4.0.7
pillow>=10.2.0
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
python-dotenv==1.0.0
whitenoise==6.6.0 
dj-database-url 
python-dotenv 
gunicorn 
gunicorn 
dj-database-url 
whitenoise 
python-dotenv 
django-cors-headers 
